requirements.txt
streamlit
opencv-python
numpy
pillow 